requirements.txt
alpaca-py==0.43.2
annotated-types==0.7.0
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
idna==3.11
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
msgpack==1.1.2
narwhals==2.17.0
numpy==2.4.1
nvidia-nccl-cu12==2.29.3
packaging==26.0
pandas==3.0.0
pillow==12.1.0
plotly==6.5.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
PyQt6==6.10.2
PyQt6-Qt6==6.10.1
PyQt6_sip==13.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
sseclient-py==1.9.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
xgboost==3.2.0
